more on types and casts
git-svn-id: https://svn.r-project.org/R/trunk@59097 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/coerce.c
@@ -161,7 +161,7 @@ IntegerFromString(SEXP x, int *warn)
 		return INT_MIN;
 	    }
 	    else
-		return xdouble;
+		return (int) xdouble;
 	}
 	else *warn |= WARN_NA;
     }
