+++ src/main/raw.c
@@ -364,7 +364,7 @@ SEXP attribute_hidden do_intToUtf8(SEXP call, SEXP op, SEXP args, SEXP env)
 	    len += used;
 	}
 	PROTECT(ans = allocVector(STRSXP, 1));
-	SET_STRING_ELT(ans, 0, mkCharLenCE(tmp, len, CE_UTF8));
+	SET_STRING_ELT(ans, 0, mkCharLenCE(tmp, (int)len, CE_UTF8));
 	if(len >= 10000) Free(tmp);
     }
     UNPROTECT(2);
