+++ src/nmath/fround.c
@@ -90,13 +90,13 @@ double fround(double x, double digits) {
     } else
 	sgn = 1.;
     if (dig == 0) {
-	return sgn * R_rint(x);
+	return (double)(sgn * R_rint(x));
     } else if (dig > 0) {
         pow10 = R_pow_di(10., dig);
 	intx = floor(x);
-	return sgn * (intx + R_rint((x-intx) * pow10) / pow10);
+	return (double)(sgn * (intx + R_rint((x-intx) * pow10) / pow10));
     } else {
         pow10 = R_pow_di(10., -dig);
-        return sgn * R_rint(x/pow10) * pow10;
+        return (double)(sgn * R_rint(x/pow10) * pow10);
     }
 }
