+++ src/nmath/rbinom.c
@@ -54,7 +54,7 @@ double rbinom(double nin, double pp)
 
     double f, f1, f2, u, v, w, w2, x, x1, x2, z, z2;
     double p, q, np, g, r, al, alv, amaxp, ffm, ynorm;
-    int i,ix,k, n;
+    int i, ix, k, n;
 
     if (!R_FINITE(nin)) ML_ERR_return_NAN;
     r = floor(nin + 0.5);
@@ -70,7 +70,7 @@ double rbinom(double nin, double pp)
 			and r == INT_MAX gave only even values */
 	return qbinom(unif_rand(), r, pp, /*lower_tail*/ 0, /*log_p*/ 0);
     /* else */
-    n = r;
+    n = (int) r;
 
     p = fmin2(pp, 1. - pp);
     q = 1. - p;
@@ -92,7 +92,7 @@ double rbinom(double nin, double pp)
 	    goto L_np_small;
 	} else {
 	    ffm = np + p;
-	    m = ffm;
+	    m = (int) ffm;
 	    fm = m;
 	    npq = np * q;
 	    p1 = (int)(2.195 * sqrt(npq) - 4.6 * q) + 0.5;
@@ -119,7 +119,7 @@ double rbinom(double nin, double pp)
       v = unif_rand();
       /* triangular region */
       if (u <= p1) {
-	  ix = xm - p1 * v + u;
+	  ix = (int)(xm - p1 * v + u);
 	  goto finis;
       }
       /* parallelogram region */
@@ -128,15 +128,15 @@ double rbinom(double nin, double pp)
 	  v = v * c + 1.0 - fabs(xm - x) / p1;
 	  if (v > 1.0 || v <= 0.)
 	      continue;
-	  ix = x;
+	  ix = (int) x;
       } else {
 	  if (u > p3) {	/* right tail */
-	      ix = xr - log(v) / xlr;
+	      ix = (int)(xr - log(v) / xlr);
 	      if (ix > n)
 		  continue;
 	      v = v * (u - p3) * xlr;
 	  } else {/* left tail */
-	      ix = xl + log(v) / xll;
+	      ix = (int)(xl + log(v) / xll);
 	      if (ix < 0)
 		  continue;
 	      v = v * (u - p2) * xll;
